refactoring into sub components
